azure_auth_series/08_production/audit-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
httpx==0.27.2
python-dotenv==1.0.1
azure-monitor-opentelemetry==1.6.4
